ModIncidentes logicaIncidentes
lógica para que el GestorIncidentes pueda notificar a la clase Incidente sobre el paquete afectado y su estado.
