+++ PROYECTO_ENCOMIENDA/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\migue\\AppData\\Roaming\\NetBeans\\20\\build.properties
+user.properties.file=C:\\Users\\Issac\\AppData\\Roaming\\NetBeans\\20\\build.properties
