+++ PROYECTO_ENCOMIENDA/src/proyecto_encomienda/GESTION_PAQUETES/BACKEND/Inventario.java
@@ -38,4 +38,10 @@ public class Inventario {
         }
         return null;
     }
+//    public void actualizarEstadoPaquete(String codigoTracking, String nuevoEstado) {
+//        Paquete paquete = obtenerPaquete(codigoTracking);
+//        if (paquete != null) {
+//            paquete.cambiarEstado(nuevoEstado);
+//        }
+//    }
 }
